[UPDATE] Add non-blocking queue
In this project, there are only 2 threads sharing the same queue.
One is for capturing packets (writing) and the other is for analyzing
and forwarding (reading). So, it's more easier to implement non-blocking
than multiple writers and readers.

Apply _sync_bool_comapre_and_swap() in  MyWrite() and Myread().
It's a function that supports atomic operation.

To reduce the critical section, using a dummy node in queue.
q->head and q->tail both point to dummy node at initialization time.
q->tail will shift to the real tail of the queue.
q->head will alway point to the dummy node.
So the real head is q->head->next.
And the point shifting operation will be done by CAS.
By this, we can make sure all cirtical section will be implemented atomically.

Signed-off-by: yenpc <asd757817@gmail.com>